ANY-52: Implement option/choice list field
