+++ package.json
@@ -62,7 +62,9 @@
     "express-winston": "^2.5.0",
     "glob": "^7.1.2",
     "istanbul": "^0.4.5",
+    "js2xmlparser": "^3.0.0",
     "jsonfile": "^4.0.0",
+    "jsontoxml": "^1.0.0",
     "jsonwebtoken": "^8.2.0",
     "lodash": "^4.17.5",
     "mongoose": "^5.0.15",
@@ -72,14 +74,14 @@
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
     "q": "^1.5.1",
-    "winston": "^2.4.1"
+    "winston": "^2.4.1",
+    "xml": "^1.0.1"
   },
   "devDependencies": {
     "chai": "^4.1.2",
     "mocha": "^5.1.0",
     "nyc": "^11.6.0",
     "objectid": "^3.2.1",
-    "selenium-webdriver": "^4.0.0-alpha.1",
     "uuid": "^3.2.1"
   }
 }
