+++ resources/platform/models/p_module.json
@@ -28,6 +28,7 @@
 	  "type": "string"
 	},
 	{
+	  "label": "Collection",
 	  "name": "ref_collection",
 	  "type": "reference"
 	},
