+++ resources/platform/models/p_option.json
@@ -3,6 +3,7 @@
   "label": "Option",
   "fields": [
     {
+      "label": "Collection",
       "name": "ref_collection",
       "type": "reference"
     },
