+++ resources/platform/models/p_rest_api.json
@@ -13,7 +13,36 @@
 	},
 	{
 	  "name": "method",
-	  "type": "option"
+	  "type": "option",
+	  "options": [
+		{
+		  "order": "100",
+		  "language": "en",
+		  "label": "GET",
+		  "value": "GET"
+		},
+		{
+		  "order": "200",
+		  "language": "en",
+		  "label": "POST",
+		  "field": "method",
+		  "value": "POST"
+		},
+		{
+		  "order": "300",
+		  "language": "en",
+		  "label": "PUT",
+		  "field": "method",
+		  "value": "PUT"
+		},
+		{
+		  "order": "400",
+		  "language": "en",
+		  "label": "DELETE",
+		  "field": "method",
+		  "value": "DELETE"
+		}
+	  ]
 	},
 	{
 	  "name": "script",
