+++ resources/platform/updates/admin-user.json
@@ -1,0 +1,10 @@
+{
+   "collection":"p_user",
+   "records":[
+      {
+         "id":"5b37598359a01c43346f8632",
+         "username": "admin",
+         "password" : "$2a$08$A3a5qp9pFtEh7bMbzonr4.R/gsWepkCAjGYhLENKFnLprCh3U8QK."
+      }
+   ]
+}
