+++ src/config/logger.js
@@ -1,5 +1,6 @@
 const winston = require('winston');
 const config = require('./config');
+const _ = require('lodash');
 
 //
 // Logging levels
@@ -35,5 +36,16 @@ const logger = new (winston.Logger)({
 });
 
 logger.options = options;
+
+let size = 50;
+
+function logBox(str) {
+  if (str)
+    console.log('* ' + _.padEnd(str, size, ' ') + ' *');
+  else
+    console.log(_.padStart('', size + 4, '*'));
+}
+
+logger.logBox = logBox;
 
 module.exports = logger;
