+++ views/styles/scss/base.scss
@@ -169,4 +169,17 @@ body {
 	  }
 	}
   }
+
+  .rotate-45-left {
+	display: inline-block;
+	-webkit-transform: rotate(-45deg);
+	transform: rotate(-45deg);
+  }
+
+  .rotate-45-right {
+	display: inline-block;
+	-webkit-transform: rotate(45deg);
+	transform: rotate(45deg);
+  }
+
 }
